requirements.txt
Babel==2.4.0
cairocffi==0.8.0
CairoSVG==1.0.22
cffi==1.10.0
cssselect==1.0.1
hjson==3.0.0
html5lib==0.999999999
Jinja2==2.9.6
lxml==3.8.0
MarkupSafe==1.0
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
pycparser==2.18
Pyphen==0.9.4
pytz==2017.2
six==1.10.0
tinycss2==0.5
WeasyPrint==0.39
webencodings==0.5.1
